src/storechat/requirements.txt
strands-agents[otel]
fastapi
uvicorn[standard]
httpx
grpcio
grpcio-tools
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-redis
kafka-python
